+++ CHANGELOG.md
@@ -40,7 +40,7 @@ NB: The login has changed to allow more configurable user identity and other att
 - Load config overrides from system properties and env (#2917)
 - Application draft can now be saved even if there are validation warnings. (#2766)
 - New application page no longer displays "Application: Success" message. (#2838)
-- Blacklist API now correctly returns 404 if user or resource do not exist when adding or removing blacklist entry. (#2835)
+- Blacklist API now returns HTTP 422 status if user or resource does not exist when adding or removing blacklist entry. (#2835)
 
 ## v2.25 "Meripuistotie" 2022-02-15